add moving logic of walls and map bounds for character

+++ src/main/kotlin/graphics/Person.kt
@@ -2,5 +2,5 @@ package graphics
 
 import java.awt.geom.Ellipse2D
 
-class Person(x: kotlin.Double,y: kotlin.Double,w: kotlin.Double, h: kotlin.Double)
-    : Ellipse2D.Double(x, y, w, h) { }
+class Person(x: kotlin.Int, y: kotlin.Int, w: kotlin.Double, h: kotlin.Double)
+    : Ellipse2D.Double(x.toDouble(), y.toDouble(), w, h) { }
